mako: touch: define a power source for charging info ported by slukk from mako-3.4-jb-mr2
Define a touch power source and add it as a
supplicant for usb, dc and wireless power suppliers
so that the external_power_changed callback is
called when there is a power supply change.

When there is a power source supplied, touch driver
must set a special bit in the touch controller
so that the touch firmware can use advanced algorithm
to minimize the noise affection.